requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.0
pandas>=2.1.0
openpyxl>=3.1.0
python-telegram-bot>=20.7
PyYAML>=6.0
aiohttp>=3.9.0
fake-useragent>=1.4.0
